split tslint rules for dev and ci
In development, console.log's are useful for debugging. This updates the
main tslint.json file to not error when console.log's are present, but
to just emit warnings. Since tslint doesn't have a stricter mode where
it fails on warnings also, I've created another tslint.ci.json file that
will emit errors if console.log's are present. This file will be used in
the CI and pre-commit hooks so when a developer is ready to commit a
change they will need to remove their debug statements.