disables the service worker
Disable the service worker as it currently loads index.html from cache for
all unknown paths. This doesn't work well with our Ingress setup because if
you visit /kubeless/ you end up getting the Dashboard instead of the Kubeapps
UI. We can consider re-enabling this once the Kubeless UI is integrated in
the Dashboard and we no longer need the external link.